Remove REGISTER_STATIC_PLUGIN_V2
This was never officially part of the public API. Remove it from
public headers so that it is not accidentally used.

This macro cannot be used with the new Qt plugins.

There are 2 unit tests that use this mechanism and for this reason,
loading of static, legacy plugins will continue but they will use
internal logic to achieve this rather than a macro defined in a
public header.

Change-Id: I3fadcc1ca5c22ea0fc3660584ddba28271d18a89
Reviewed-by: Lorn Potter <lorn.potter@nokia.com>