Changes so that sensor.start() does not block (QML)
The C++ library tries to delay things as much as possible. It turns
out that doing this for QML is not optimal because it means calling
sensor.start() can block for a long time (up to hundreds of
milliseconds in worst case scenarious). If the sensor is started
after the UI is visible, this will cause a stutter.

The workaround is not particularly elegant. We force the sensors
library to do its setup while the QML plugin is being loaded.

Change-Id: Iae945ccb79eb568c5decbfdf8a5d623cd96e9d6b
Reviewed-by: Zsolt Simon <zsolt.simon@nokia.com>