setDataRates clobbers user-set data rate
Under normal conditions, dataRate is set/checked like so.

QSensor::setDataRate records the data rate. It can't be evaluated because we have
no backend. Upon getting a backend, we are notified of the available data rates.
Re-evaluate the previously-set dataRate after connecting to a backend.

The problem is that connectToBackend did not save the dataRate value until after
setDataRates was called and setDataRates was clobbering the dataRate property.

Two fixes will be made. connectToBackend will save the dataRate before creating
the backend and setDataRates will stop clobbering dataRate.

Also, setDataRates is now banned from being called outside of the constructor
because otherwise an available data rate may disappear during use.

Add a test case to verify that a previously-set data rate is not clobbered.

Change-Id: I9632b2c8ffcc6d48d9bc91f263b99028ee997ff9
Reviewed-by: Lorn Potter <lorn.potter@nokia.com>