Introduce QRotationReading::setFromEuler() in favor of set{X/Y/Z}()
In preparation of adding new representations to the reading (QTBUG-25840),
the setX(), setY() and setZ() setters were removed, as these don't set
the representation atomically, and don't provide a single point where
a  conversion to matrix or quaternion can occur. These setters were
only supposed to be used by the backends, and all backends are
ported in this commit.

Change-Id: Ib652520578d293687fc8515d226f1f61aa2f2def
Reviewed-by: Lincoln Ramsay <lincoln.ramsay@nokia.com>
Reviewed-by: Lorn Potter <lorn.potter@nokia.com>