iOS: change accelerometer to use QTimer for polling
According to the docs (CMMotionManager class reference) there are
two ways of interacting with the accelerometer; Either through the
callback API (NSOperationQueue), or peridic sampling (polling).
Our first implementation of IOSAcceleometer used the former
technique, which turns out to have bad performance when using the
sensor together with a fine-grained QTimer. And this case is
pretty common when using sensors together with e.g QML.

Reading through the docs more carefully, they recommend using the
polling technique when creating games instead since the
NSOperationQueue introduces some overhead. So this patch does
that, change the implementation to use QTimer based polling. And
this solves the performance issues found.

Change-Id: Ifde0d2292302467afb8db90a954ef45f3238350e
Reviewed-by: Thomas McGuire <thomas.mcguire@kdab.com>