Define the sensor coordinate system unambiguously
All illustrations in the documentation show devices with phone
form factors, i.e. portrait orientation. Some tablets have a native
landscape orientation. It is not clear from the documentation
if the sensor coordinate system should be relative to the portrait
orientation or the native orientation.

This change defines the native orientation as the canonical orientation.

Task-number: QTBUG-32144
Change-Id: I870cb3c9b5aec04f2ed8394a4407b53b888e82c3
Reviewed-by: Thomas McGuire <thomas.mcguire@kdab.com>