Implement QAccelerometer accelerationMode
The Qt API does not expose Linear Accel and Gravity as separate sensor types,
they are different versions of QAccelerometer, accessed by modifying the
accelerationMode property. The possible values are Combined, User, and Gravity, which
correspond to Acceleration, Linear Acceleration, and Gravity respectively. When
constructing the accelerometer backend, the accelerationMode must be queried. The
backend must also connect to the accelerationModeChanged signal from the QAccelerometer.

I think that the existing implementation of QSensors on Android assumes that there is a
one-to-one correspondence between QSensors and AndroidSensorTypes. To fix this assumption,
I promoted the m_type attribute of androidcommonsensor.h from private to protected so
that a sensor backend can update the Android sensor being used.

Task-number: QTBUG-30531
Change-Id: I842d2485109542f74731d179e2720ea55fa248cc
Reviewed-by: Fabian Bumberger <fbumberger@rim.com>
Reviewed-by: Lorn Potter <lorn.potter@jollamobile.com>