Create a new sensor type for distance.
Create a new sensor type for distance. This sensor type supports new hardware sensors
that can measure physical distance from the device, in centimeters. The API is designed
to mimic the Android API for proximity (TYPE_PROXIMITY), so that if a given proximity
sensor only supports a binary measurement (near vs far), then instead of reporting distance
in cm, the QDistanceSensor will return the max range value to represent far, and a lesser
value to represent close. Using this definition should simplify implementation.

The main reason for not implementing this as a new property of QProximitySensor is that
clients of QProximitySensor have made the assumption that they will receive the readingReady
signal if and only if the reading has changed from near to far or vice versa. Adding a
distance property will break that assumption, as distance has a higher degree of precision.

Change-Id: Ia804948c78ff7391fc8b78df975cddcf861326dc
Reviewed-by: Fabian Bumberger <fbumberger@rim.com>
Reviewed-by: Bernd Weimer <bweimer@blackberry.com>
Reviewed-by: Lorn Potter <lorn.potter@jollamobile.com>