Fix a couple of issues in the maze example
1.) Remove dependency on QtSystemInfo. The module is not officially
supported at this stage and screen saver related code is not really
required for the example.

2.) Add AndroidManifest.xml to lock the screen orientation in portrait
mode. Since the example depends on the tilt sensor a rotating UI is
messing the game up.

3.) Port app to current TiltSensor API.

The game works but its biggest problem remains the UI. It uses
fixed-size UI elements which makes it tiny on High DPI screens.

Task-number: QTBUG-31076
Change-Id: I73df9361b5171da930df66072e8ee9eba89b71ad
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>