Make GenericTiltSensor::calibrate known to QMetaObject
When using qml calibrating the tilt sensor does not work on WinRT and
Windows Phone without this change.
QmlTiltsensor::calibrate calls QTiltSensor::calibrate which calls
QMetaObject::invokeMethod with "calibrate" on its backend. That causes
the message "No such method QSensorBackend::calibrate()" without this
change. Thus the meta system has to be made aware of that function.

Change-Id: I539c0fb44e20fffb78bf515ba3767dafa3ce4ed6
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@theqtcompany.com>