iOS: protect OS sensors from being stopped when shared by several Qt sensors
When a Qt sensor is told to stop, it automatically stops the underlying
OS sensor as well. This caused a problem when several Qt sensors were
running, since stopping one would stop them all.

This patch will add a ref count for each affected sensor, so that we
keep track of when it's safe to start and stop the underlying OS sensor.

Task-number: QTBUG-54977
Change-Id: If06c4daac22916feef2f4c7bd521363d74f342a6
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Lorn Potter <lorn.potter@canonical.com>