Be consistent about values related to pi (radian / degree ratio)
Use JavaScript's Math.PI to compute 180 / pi (the number of degrees in
one radian), rather than a hand-coded value for this constant.  Store
the result in a readonly property double to avoid repeating the
computation.  Tweak the pitch and roll calculations to use more apt
Math methods: use hypot rather than sqrt of a sum of squares, and
use atan2 instead of a tan of a ratio (whose denominator is always
positive, so this doesn't change which angle we get).

Task-number: QTBUG-58083
Change-Id: Ic3f865dfcbd9b02a0b93f065995d850386aef18c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>