testSingleGestures: Fix for case of multiple gestures
This test fails sometimes for the twist gesture data, because
it contains 3 gestures. QTRY_COMPARE_WITH_TIMEOUT usually stops
comparison after recognition of the first gesture, but it may
make the first comparison after more than one gesture has been
recognized and thus fail the test.

Add a new column to the test data with a list of gestures that
should be recognized and use the list in QTRY_COMPARE_WITH_TIMEOUT.
It will make the error message more informative if the test will
fail, and should make the test pass on all configurations.

Change-Id: Icfd86831b527da4933da2a20f45ccf9617862eac
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>