Update AndroidManifest in accelbubble example
This example has an AndroidManifest.xml in order to force the
orientation to portrait, but it had become outdated, setting
the target version to android-14, which has not been supported
since Qt 5.7. This updates the manifest to the latest template,
while also still setting the orientation to portrait.

[ChangeLog][Android][Examples] Updated AndroidManifest.xml in
accelbubble example to be compatible with Qt again.

Task-number: QTBUG-70159
Change-Id: I80a34c34837682c3ff2a7d232d527820fb9f4ca4
Reviewed-by: BogDan Vatra <bogdan@kdab.com>