Say hello to Qt Android Sensors based on NDK API
Until Qt 5.9, we had to use JNI way because the NDK sensors API was
introduced in API-16. Starting with Qt 5.9 we can safely use the NDK API
as API_16 is the minimum API supported by Qt.
Using the NDK API instead of JNI boost the performance a lot! The CPU
usage dropped 2 to 4 times, from over 60% using the JNI to less than
20%.

[ChangeLog][Android] Rewrite Qt Android Sensors using NDK
API.

Change-Id: If3e3f1e56973e2a189662912b2a6ebfbd71dee14
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>