Update commercial license headers
Updated header.COMM to the files in tqtc-qtsensors. Examples,
tests or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4213
Change-Id: I0fb0c2dbfe22d2df2f55eaeb5e1cdbddb98afa71
Reviewed-by: Antti Kokko <antti.kokko@qt.io>