Enable QtSensors manual tests as part of build
The two manual test applications are now part of build when manual
tests are built. This helps to detect if the test applications break
over time.

The tests were slightly renamed to avoid a naming clash with an
example app. In addition the commit addresses few trivial things
so that the applications start and run without warnings.

Task-number: QTBUG-92502
Change-Id: Ia0368764102efc16276116bc8c75d562d79c1e27
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>