Fix and simplify sensor_explorer example
This commit strives to both fix and simplify the application
and bring its intended demonstrative purpose more forward,
consequently hopefully also making the example easier to
understand and maintain.

First, the sensor_explorer example used QQC1 functionalities
which are no longer available in Qt6. The removed elements were
such that the datamodels needed to be redesigned. This commit
changes the example to use bit more standard Qt model-view
approach.

Second, the sensor_explorer application has not worked in years
on Android due to missing QML plugin installation. While this
could be fixed, in the name of simplicity the example is
now a single binary.

Third, the explorer allowed editing of sensor values. While
this could be a feature of the example, it perhaps deviates
from the gist of the example making it more difficult to
understand. Thus the editing was not added as part of the
example rewrite. It can be added later though.

Task-number: QTBUG-92514
Task-number: QTBUG-92505
Task-number: QTBUG-63054
Change-Id: Icd620ea605a87ba74a42cd1309a3c02d05b146ac
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>