Fix tst_qsensor autotestcase to pass on android device
The main issue with running the autotests on an Android device is
that the device has real sensors too and the test case is not prepared
for that. This commit addresses this by allowing additional available
sensors as well as specifying more specifically which sensor IDs to use
for testing.

Task-number: QTBUG-73617
Task-number: QTBUG-92513
Change-Id: I92180df6f5e8361d39134567616b21a3a51b636c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>