Make sensor possible to indicate it is no longer busy
The current implementation allowed sensor backend to only indicate
that it is busy, but not to tell when it was freed again. This commit
allows sensor backend to indicate also that it's busy state has cleared.

It is up to the sensor implementation to decide if it makes sense /
is possible to do that or not.

Task-number: QTBUG-92513
Task-number: QTBUG-92505
Change-Id: Ied4857850e81346031fd83aa347d9955081118e8
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit f8445fdcbf75e455443304bc290c48c37961e9f1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>