Fix tst_sensorgestures_gestures to pass on Android
The problem was that the testcase did not find the testdata files on
Android. On closer inspection it turns out that TESTDATA token in
qt_internal_add_test treats ANDROID/IOS and rest of the platforms
differently in terms of how test resources are handled. This commit
stores testdata unconditionally as qrc resources.

Pick-to: 6.2
Task-number: QTBUG-92513
Task-number: QTBUG-73618
Change-Id: I0cfe878661dd2c55fa63abc1f95534a8041aa34d
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>