Remove TI Sensor tag support
The TI Sensor tag support is specific to a firmware version, and while
it could be updated on an ongoing basis, there has not been expressed
use for it beyond its initial use cases.

As with Qt sensors in general, sensors and platforms can be later
reintroduced without compatibility issues should such needs emerge.

[ChangeLog][TI Sensor tag] Remove support

Pick-to: 6.2
Task-number: QTBUG-92512
Change-Id: I70c8514701787cc1d21195546a48b38d8009434f
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>