Update commercial license headers
Updated header.COMM to the files in tqtc-qtsensors.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4957
Change-Id: Id541ba2aeccb375245540eea0516f566f1aee280
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>