Remove disabled sensor gesture code
Support for sensor gestures in their then-current form was removed
in Qt 6 (QTBUG-95747). This commit removes the disabled code. This is
to avoid distraction when keeping the repository up-to-date with rest of
Qt.

The gestures may be reintroduced in some shape or form in the future
(QTBUG-97066). For those reference purposes the old code can be found
in the git history.

Fixes: QTBUG-107065
Change-Id: If2c50171f84d483dde55600422d138d550124bdc
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>