Remove sensors maze example
The Maze (QML) example is a nice game implementation, but doesn't add
much demonstration value in comparison to sensorsshowcase example.

The one unique thing it does demonstrate is tilt sensor. Tiltsensor can
be added to the sensorsshowcase too; but OTOH tilt is in some sense
just 'a' sensor among the sensors, and there is probably no need to use
all of them in the examples.

Task-number: QTBUG-110939
Pick-to: 6.5
Change-Id: I2fd997750b05b3962428955a2929e149af6b72bd
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>