Remove sensors qmlqtsensors example
The example doesn't add much demonstration value in comparison to
sensorsshowcase example.

The one unique thing it does demonstrate is ambient light sensor.
That sensor can be added to the sensorsshowcase too; but OTOH it is in
some sense just 'a' sensor among the sensors, and there is probably no
need to use all of them in the examples.

Task-number: QTBUG-110939
Change-Id: I5d0ba59d969b532689e23388cbdda42884838fca
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>
(cherry picked from commit 849519a842cf34527f4532837dcf84ec6bb28108)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>