Remove sensors grue example
The primary purpose of the grue example is to demonstrate how to create
new sensors. However, while a nice implementation, it doesn't achieve
this purpose particularly well, and  is all in all an overly complex
construct given the straightforward purpose (for instance causing build
and deployment challenges).

Instead it should be replaced with a tutorial-like step-by-step guide
on how to add new sensors. The item to track this: QTBUG-111251

Task-number: QTBUG-110939
Pick-to: 6.5
Change-Id: If353fd019483ad3d608f16d90820c09e676f1df1
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>