Move sensor_explorer QML example to manual tests
The example is useful as a sensor testing tool, but less necessary
as an example. The primary demonstrative element of the example was
how to iterate the available sensors, which is covered in a
documentation snippet (followup commit).

This commit also renames the pre-existing widget-based sensor_explorer
manual test to avoid name conflicts, as well as removes the qdoc and qmake
support from the now-a-manual-test application.

Task-number: QTBUG-110939
Pick-to: 6.5
Change-Id: I422f62f852d0a7e40a76f555ec8aa98404164f7a
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>