Doc: Add DocBook equivalent for custom table
QDoc provides a "\raw" command that allows a user to inject
format-specific code that complements the code that QDoc generates.

The "\raw" command expect the user to provide a "format" argument to
ensure that the provided code is generated only for the format that is
written for.

Due to a bug in QDoc DocBook generation, the DocBook documentation was
generating, sometimes incorrectly, "\raw" commands intended for the HTML
format.

The Qt Sensors documentation uses the "\raw HTML" commands to
provide some custom-colored table.

Due to the aforementioned bug, QDoc was generating some equivalent code
for DocBook, albeit no user-provided DocBook equivalent existed.

As the bug is now being fixed, the DocBook documentation for Sensors
will lose the custom-colored table in its output.

To preserve the custom-colored table when generating the DocBook
documentation, an equivalent "\raw DocBook" command is now provided.

Change-Id: I249dec021719d1f7823e813b67e4de5ce93b49a6
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>