Doc: List of all Qt overviews which are now termed as explanation
The autogenerated list of overviews was adding the \group command which
included all the groups instead of overviews.
The idea here is to categorize the overviews later on once we have
the list of all overviews.

Task-number: QTBUG-114762
Change-Id: Iab8b6c00f38f0741e200756a178b901f1d195c4f
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
(cherry picked from commit ecacbe6f4e0e598e6ff25c521959384932252c04)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>